+++ pom.xml
@@ -67,6 +67,7 @@
         <shade-plugin.version>2.3</shade-plugin.version>
         <nexus-plugin.version>1.6.5</nexus-plugin.version>
         <ghp-plugin.version>1.0.0</ghp-plugin.version>
+        <dropwizard.version>0.9.0</dropwizard.version>
     </properties>
 
     <dependencies>
@@ -94,6 +95,11 @@
             <groupId>se.walkercrou</groupId>
             <artifactId>google-places-api-java</artifactId>
             <version>2.1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>io.dropwizard</groupId>
+            <artifactId>dropwizard-core</artifactId>
+            <version>${dropwizard.version}</version>
         </dependency>
     </dependencies>
