+++ property-analytics.yaml
@@ -1,0 +1,2 @@
+http:
+  rootPath: /api/*
