+++ src/main/java/com/di/analytics/PropertyAnalyticsConfiguration.java
@@ -1,0 +1,9 @@
+package com.di.analytics;
+
+import io.dropwizard.Configuration;
+
+/**
+ * Created by hughsoong on 10/30/2015.
+ */
+public class PropertyAnalyticsConfiguration extends Configuration {
+}
