+++ src/main/java/com/di/analytics/PropertyAnalyticsResource.java
@@ -1,0 +1,14 @@
+package com.di.analytics;
+
+
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+
+/**
+ * Created by hughsoong on 10/30/2015.
+ */
+@Path("/property")
+@Produces(MediaType.APPLICATION_JSON)
+public class PropertyAnalyticsResource {
+}
