+++ src/main/java/com/di/analytics/PropertyAnalyticsResults.java
@@ -1,0 +1,10 @@
+package com.di.analytics;
+
+/**
+ * Created by hughsoong on 10/30/2015.
+ */
+public class PropertyAnalyticsResults {
+    public PropertyAnalyticsResults(){
+        //constructor
+    }
+}
