+++ src/main/resources/places_api.key
@@ -1,1 +1,1 @@
-${env.GOOGLE_PLACES_API_KEY}
+AIzaSyDaihne89yFZ7qWZcvIa1uXLi94oJ24Ux4
